mali: utgard: Keep mali_internal_sync_fence structure synced w/ sync_file
In the change 74fa0af47306 (BACKPORT: dma-buf/sync_file: Allow
multiple sync_files to wrap a single dma-fence), an extra flag
value was added to the struct sync_file.

This causes trouble with the mali driver, as it has its own
mali_internal_sync structure that tries to mirror the sync_file.

Thus this patch tries to fix up the mali_internal_sync_fence to
match, at least for 4.9 kernels.

Test: before change, crashes during init, after change, is ok
Change-Id: Icdf7f175b3df9a0c316e527aca66680792cd08e8
Signed-off-by: John Stultz <john.stultz@linaro.org>