drm: adv7511: Avoid setting DRM_MODE_TYPE_PREFERRED on all valid modes
For HiKey/HiKey960 we have a mode-validation whitelist hack here
in the adv7511 bridge code, as the endcoder/crtc doesn't yet
support mode_valid callbacks in 4.9.

In this hack, we incorrectly set all valid modes as
DRM_MODE_TYPE_PREFERRED. There should be at most one PREFERRED
mode, which causes drm_hwcomposer trouble picking the best
mode as instead of picking the first PREFERRED mode, it picks
the last.

This results in a very low-res mode being selected on HiKey.

To fix this, don't set any mode as PREFERRED. The extra state
tracking to pick a single PREFERRED mode would be extra
complexity in a hack, and if no modes are PREFERRED the first
mode is chosen, which is what we want anyway.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Iff58856a53d5da8d64d006cc3584b25dae464dad